Revert "RESTRICT AUTOMERGE Set the container field of aconfig flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a002618d-3968-4b19-961a-12aa91d6fc8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5cfaece771578bb2c26b7baa64ed52a0e5430630