Revert "RESTRICT AUTOMERGE Set the container field of aconfig flags"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e7784154-82a7-4f66-bad2-ffc546d237d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If6b2547361f8afcfeb1e3a09b28d5ab61d0375f9